add BoardFilterSelector
