+++ src/components/Organisms/Board/MemberOtherBoard.tsx
@@ -1,6 +1,7 @@
+import BoardFilterSelector from '@src/components/Molecules/Board/BoardFilterSelector';
 import OtherBoardMember from '@src/components/Molecules/Board/OtherBoardMember';
 import { useOtherBoard } from '@src/hooks/useBoards';
-import React, { useMemo } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import BoardListItem from './BoardListItem';
@@ -9,12 +10,17 @@ import BoardListItem from './BoardListItem';
 const LIMIT = 1000;
 
 const MemberOtherBoard = () => {
+  const [filterBoard, setFilterBoard] = useState<string>('최신순');
   const { memberId } = useParams();
   const { data: boardList } = useOtherBoard({
     limit: LIMIT,
     memberId:
       typeof memberId === 'string' ? parseInt(memberId) : memberId ?? null,
   });
+
+  const onChangeFilter = useCallback((filterBoard: string) => {
+    setFilterBoard(filterBoard);
+  }, []);
 
   const title = useMemo(() => {
     if (boardList) {
@@ -23,12 +29,23 @@ const MemberOtherBoard = () => {
     return `작성자의 다른 글 (${0})`;
   }, [boardList]);
 
+  console.log(boardList?.boardOfMemberResDtoList);
   const list = useMemo(() => {
     if (boardList) {
-      return boardList.boardOfMemberResDtoList;
+      if (filterBoard === '조회순') {
+        return boardList.boardOfMemberResDtoList.sort((a, b) => {
+          return b.boardCnt ?? 0 - (a.boardCnt ?? 0);
+        });
+      } else if (filterBoard === '좋아요순') {
+        return boardList.boardOfMemberResDtoList.sort((a, b) => {
+          return b.heartCnt ?? 0 - (a.heartCnt ?? 0);
+        });
+      } else {
+        return boardList.boardOfMemberResDtoList;
+      }
     }
     return [];
-  }, [boardList]);
+  }, [boardList, filterBoard]);
 
   return (
     <Container>
@@ -42,10 +59,17 @@ const MemberOtherBoard = () => {
         />
       ) : null}
       <Wrapper>
-        <OtherBoardTitle>{title}</OtherBoardTitle>
+        <Header>
+          <OtherBoardTitle>{title}</OtherBoardTitle>
+          <BoardFilterSelector
+            active={true}
+            filter={filterBoard}
+            onChangeFilter={onChangeFilter}
+          />
+        </Header>
 
         {list.map((board) => (
-          <BoardListItem board={board} />
+          <BoardListItem board={board} key={board.boardId} />
         ))}
       </Wrapper>
     </Container>
@@ -59,6 +83,14 @@ const Container = styled.div`
   height: 100%;
   flex-direction: column;
   flex: 1;
+`;
+
+const Header = styled.div`
+  display: flex;
+  flex-direction: row;
+
+  justify-content: space-between;
+  width: 100%;
 `;
 
 const Wrapper = styled.div`
